// generated/hypertune.ts
export const vercelFlagDefinitions = {"ctaVariant":{"options":[{"label":"Off","value":false},{"label":"On","value":true}],"origin":"https://app.hypertune.com/projects/7808/main/draft/logic?selected_field_path=root%3EctaVariant"},"exampleFlag":{"options":[{"label":"Off","value":false},{"label":"On","value":true}],"origin":"https://app.hypertune.com/projects/7808/main/draft/logic?selected_field_path=root%3EexampleFlag"}};

export type FlagValues = {
  "ctaVariant": boolean;
  "exampleFlag": boolean;
}

export type FlagPath = keyof FlagValues & string;

export const flagFallbacks: FlagValues = {
  "ctaVariant": false,
  "exampleFlag": false,
}

export function decodeFlagValues<TFlagPath extends keyof FlagValues & string>(
  encodedFlagValues: string,
  flagPaths: TFlagPath[]
): Pick<FlagValues, TFlagPath> {
  return sdk.decodeFlagValues({ encodedFlagValues, flagPaths })
}

export type VariableValues = {};

export type User = {
  id: string;
  name: string;
  email: string;
}

export const EnvironmentEnumValues = [
  "development",
  "production",
  "test"
] as const;
export type Environment = typeof EnvironmentEnumValues[number];

/**
 * This `Context` input type is used for the `context` argument on your root field.
 * It contains details of the current `user` and `environment`.
 * 
 * You can define other custom input types with fields that are primitives, enums 
 * or other input types.
 */
export type Context = {
  user: User;
  environment: Environment;
}